Update and rename assignmentFunctions.js to flashcardFunctions.js
Omit lines of code that related to MongoDB such as client, db etc. Instead, I used schema and default features of mongoose
